requirements-dev.txt
-r requirements.txt
flake8==3.9.2
ipdb==0.13.7
factory-boy==3.2.0
Faker==8.1.4
pytest-cov==2.12.0
pytest-django==4.3.0
pytest-flake8==1.0.7
pytest-pythonpath==0.7.3
pytest==6.2.4
